requirements.freeze.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.109.2
h11==0.14.0
idna==3.6
iso8601==1.1.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pydantic==2.6.2
pydantic_core==2.16.3
pypika-tortoise==0.1.6
pytz==2024.1
setuptools==69.0.2
sniffio==1.3.0
starlette==0.36.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.27.1
wheel==0.42.0
